refactor: getnetwork
add to handlers and expose to wasm
