+++ pkg/wasm/context/context.go
@@ -7,6 +7,7 @@ import (
 	"context"
 
 	"github.com/rumsystem/quorum/internal/pkg/appdata"
+	"github.com/rumsystem/quorum/internal/pkg/options"
 	quorumP2P "github.com/rumsystem/quorum/internal/pkg/p2p"
 	"github.com/rumsystem/quorum/internal/pkg/storage"
 	quorumConfig "github.com/rumsystem/quorum/pkg/wasm/config"
@@ -19,15 +20,17 @@ func GetWASMContext() *QuorumWasmContext {
 	return wasmCtx
 }
 
-func Init(qchan chan struct{}, config *quorumConfig.BrowserConfig, node *quorumP2P.Node, appDb *appdata.AppDb, dbMgr *storage.DbMgr, ctx context.Context, cancel context.CancelFunc) {
-	wasmCtx = NewQuorumWasmContext(qchan, config, node, appDb, dbMgr, ctx, cancel)
+func Init(qchan chan struct{}, config *quorumConfig.BrowserConfig, node *quorumP2P.Node, ethAddr string, nodeOpt *options.NodeOptions, appDb *appdata.AppDb, dbMgr *storage.DbMgr, ctx context.Context, cancel context.CancelFunc) {
+	wasmCtx = NewQuorumWasmContext(qchan, config, node, ethAddr, nodeOpt, appDb, dbMgr, ctx, cancel)
 }
 
 type QuorumWasmContext struct {
-	QNode  *quorumP2P.Node
-	AppDb  *appdata.AppDb
-	DbMgr  *storage.DbMgr
-	Config *quorumConfig.BrowserConfig
+	EthAddr string
+	NodeOpt *options.NodeOptions
+	QNode   *quorumP2P.Node
+	AppDb   *appdata.AppDb
+	DbMgr   *storage.DbMgr
+	Config  *quorumConfig.BrowserConfig
 
 	Ctx    context.Context
 	Cancel context.CancelFunc
@@ -35,7 +38,7 @@ type QuorumWasmContext struct {
 	Qchan chan struct{}
 }
 
-func NewQuorumWasmContext(qchan chan struct{}, config *quorumConfig.BrowserConfig, node *quorumP2P.Node, appDb *appdata.AppDb, dbMgr *storage.DbMgr, ctx context.Context, cancel context.CancelFunc) *QuorumWasmContext {
-	qCtx := QuorumWasmContext{node, appDb, dbMgr, config, ctx, cancel, qchan}
+func NewQuorumWasmContext(qchan chan struct{}, config *quorumConfig.BrowserConfig, node *quorumP2P.Node, ethAddr string, nodeOpt *options.NodeOptions, appDb *appdata.AppDb, dbMgr *storage.DbMgr, ctx context.Context, cancel context.CancelFunc) *QuorumWasmContext {
+	qCtx := QuorumWasmContext{ethAddr, nodeOpt, node, appDb, dbMgr, config, ctx, cancel, qchan}
 	return &qCtx
 }
