+++ pkg/wasm/register.go
@@ -59,6 +59,20 @@ func RegisterJSFunctions() {
 		return Promisefy(handler)
 	}))
 
+	js.Global().Set("GetNetwork", js.FuncOf(func(this js.Value, args []js.Value) interface{} {
+		handler := func() (map[string]interface{}, error) {
+			ret := make(map[string]interface{})
+			res, err := quorumAPI.GetNetwork()
+			if err != nil {
+				return ret, err
+			}
+			retBytes, err := json.Marshal(res)
+			json.Unmarshal(retBytes, &ret)
+			return ret, nil
+		}
+		return Promisefy(handler)
+	}))
+
 	js.Global().Set("GetContent", js.FuncOf(func(this js.Value, args []js.Value) interface{} {
 		if len(args) < 4 {
 			return nil
